move the test file to mermaid package and exclude it in the pyproject.toml file
